feat:add RacingController
